backend/requirements.txt
-i https://pypi.org/simple
alembic==1.12.1; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; sys_platform == 'win32'
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
dominate==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.1.0.post1
flask==3.0.0
flask-bootstrap==3.3.7.1
flask-cors==4.0.0
flask-login==0.6.3
flask-mail==0.9.1
flask-marshmallow==0.15.0
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
marshmallow-sqlalchemy==0.29.0
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9
pyjwt==2.8.0
pymupdf==1.23.6
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23; python_version >= '3.7'
sqlalchemy-utils==0.41.1
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
visitor==0.1.3
werkzeug==3.0.1; python_version >= '3.8'
wtforms==3.1.1; python_version >= '3.8'
wtforms-sqlalchemy==0.3
